# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── Users ─────────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: markAttendance
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # ── Assignments ───────────────────────────────────────────────────────
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: teacherId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assignments/{id}/submissions:
    get:
      operationId: getSubmissions
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: submitAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ── Exams ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # User
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        subjectIds:
          type: array
          items:
            type: integer
        avatarUrl:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        employeeId:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Attendance:
      type: object
      required: [id, studentId, studentName, classId, date, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, excused]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [studentId, classId, date, status]
      properties:
        studentId:
          type: integer
        classId:
          type: integer
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, excused]
        notes:
          type: string
    AttendanceSummary:
      type: object
      required: [totalDays, presentDays, absentDays, lateDays, attendanceRate]
      properties:
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        excusedDays:
          type: integer
        attendanceRate:
          type: number

    # Assignment
    Assignment:
      type: object
      required: [id, title, classId, subjectId, dueDate, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        dueDate:
          type: string
        status:
          type: string
          enum: [active, closed]
        totalMarks:
          type: ["integer", "null"]
        submissionCount:
          type: integer
        createdAt:
          type: string
    AssignmentInput:
      type: object
      required: [title, classId, subjectId, teacherId, dueDate]
      properties:
        title:
          type: string
        description:
          type: string
        classId:
          type: integer
        subjectId:
          type: integer
        teacherId:
          type: integer
        dueDate:
          type: string
        totalMarks:
          type: integer
    AssignmentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [active, closed]
        totalMarks:
          type: integer

    # Submission
    Submission:
      type: object
      required: [id, assignmentId, studentId, studentName, submittedAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        fileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        marksObtained:
          type: ["integer", "null"]
        feedback:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, graded, returned]
        submittedAt:
          type: string
    SubmissionInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        fileUrl:
          type: string
        notes:
          type: string

    # Exam
...